+++ src/pythonPreview.ts
@@ -3,6 +3,7 @@ import * as path from "path";
 import * as vscode from "vscode"
 import {Limit} from "./throttle"
 import Utilities from "./utilities"
+import {settings} from "./settings"
 
 /**
  * shows AREPL output (variables, errors, timing, and stdout/stderr)
@@ -15,7 +16,6 @@ export default class PythonPreview{
     public throttledUpdate: () => void
 
     private _onDidChange: vscode.EventEmitter<vscode.Uri>;
-    private settings: vscode.WorkspaceConfiguration;
     private lastTime: number = 999999999;
 
     private html;
@@ -27,15 +27,11 @@ export default class PythonPreview{
     <p style="font-size:14px">⚠ <b style="color:red">WARNING:</b> code is evaluated WHILE YOU TYPE - don't try deleting files/folders! ⚠</p>
     <p>evaluation while you type can be turned off or adjusted in the settings</p>
     <br>
-    <h3>AREPL 1.0.13 🚀 🐛 "mononoke"</h3>
+    <h3>AREPL 1.0.14 🚀 🐛 "one punch"</h3>
     <ul>
-        <li>🚀 right click on editor title to launch arepl</li>
+        <li>🚀 Setting changes now take effect instantly (no need to reload arepl)</li>
 
-        <li>🚀 Added cache var. Use "areplStore" to store data inbetween runs! See <a href="https://github.com/Almenon/AREPL-vscode/wiki/Caching-data-between-runs">wiki</a> for futher info.</li>
-
-        <li>🐛 Fixed vars dissapearing when there is syntax error</li>
-
-        <li>🐛 Fixed vars not clearing when using gui library</li>
+        <li>🐛 Fixed silent spawn error on mac</li>
     </ul>
     <br>
     
@@ -123,7 +119,6 @@ if r.status_code == 200:
         this._onDidChange = new vscode.EventEmitter<vscode.Uri>();
         this.css = `<link rel="stylesheet" type="text/css" href="${this.getMediaPath("pythonPreview.css")}">`
         this.jsonRendererScript = `<script src="${this.getMediaPath("jsonRenderer.js")}"></script>`
-        this.settings = vscode.workspace.getConfiguration("AREPL");
 
         if(htmlUpdateFrequency != 0){
             // refreshing html too much can freeze vscode... lets avoid that
@@ -155,8 +150,8 @@ if r.status_code == 200:
             window.onload = function(){
                 ${userVarsCode}
                 let jsonRenderer = renderjson.set_icons('+', '-') // default icons look a bit wierd, overriding
-                    .set_show_to_level(${this.settings.get("show_to_level")}) 
-                    .set_max_string_length(${this.settings.get("max_string_length")});
+                    .set_show_to_level(${settings().get("show_to_level")}) 
+                    .set_max_string_length(${settings().get("max_string_length")});
                 document.getElementById("results").appendChild(jsonRenderer(userVars));
             }
             </script>`
@@ -247,8 +242,8 @@ if r.status_code == 200:
 
     private updateContent(){
 
-        const printPlacement = this.settings.get<string>("printResultPlacement")
-        const showFooter = this.settings.get<boolean>("showFooter")
+        const printPlacement = settings().get<string>("printResultPlacement")
+        const showFooter = settings().get<boolean>("showFooter")
 
         // todo: handle different themes.  check body class: https://code.visualstudio.com/updates/June_2016
         this.html = `<!doctype html>